--- kobo/apps/stripe/tests/test_utils.py
@@ -1,27 +1,0 @@
-from djstripe.models import Price, Product
-from model_bakery import baker
-
-from kobo.apps.stripe.utils import get_default_plan_name
-from kpi.tests.kpi_test_case import BaseTestCase
-
-
-class UtilsTestCase(BaseTestCase):
-    def test_get_default_plan_name(self):
-        assert get_default_plan_name() is None
-
-        product = baker.prepare(
-            Product,
-            metadata={'product_type': 'plan', 'default_free_plan': 'true'},
-            active=True,
-            name='Test Community Plan',
-        )
-        product.save()
-        baker.make(
-            Price,
-            active=True,
-            id='price_1LsSOSAR39rDI89svTKog9Hq',
-            product=product,
-            metadata={'max_purchase_quantity': '3'},
-        )
-
-        assert get_default_plan_name() == product.name
